Cleaned up excess files and renamed server
