+++ .gitignore
@@ -1,4 +1,19 @@
 node_modules
 .DS_Store
-.LSOverride
 .AppleDouble
+.LSOverride
+azure.err
+iisnode.yml
+
+# Icon must ends with two \r.
+Icon
+
+
+# Thumbnails
+._*
+
+# Files that might appear on external disk
+.Spotlight-V100
+.Trashes
+node_modules/
+client/lib
