--- doozy/.gitignore
@@ -1,18 +1,0 @@
-.DS_Store
-.AppleDouble
-.LSOverride
-azure.err
-iisnode.yml
-
-# Icon must ends with two \r.
-Icon
-
-
-# Thumbnails
-._*
-
-# Files that might appear on external disk
-.Spotlight-V100
-.Trashes
-node_modules/
-client/lib
